[FEATURE] Change expression level to median approach.

+++ include/ibf.h
@@ -120,7 +120,7 @@ void read_binary(robin_hood::unordered_node_map<uint64_t, uint16_t> & hash_table
 * \param counts               Vector, where the number of minimiser at a certain expression level should be stored into.
 */
 void read_header(arguments & args, ibf_arguments & ibf_args, std::filesystem::path filename,
-                 std::vector<uint64_t> & counts);
+                 std::vector<uint16_t> & counts);
 
 /*! \brief Calculates statistics from header files created by minimiser
  * \param args               The minimiser arguments to use (seed, shape, window size).
